requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
qrcode[pil]==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
email-validator==2.0.0
stripe==7.8.0
